requirements.txt
streamlit==1.32.0
opencv-python-headless==4.9.0.80
deepface==0.0.79
transformers==4.38.2
torch==2.2.1
numpy==1.26.4
pandas==2.2.1
pyyaml==6.0.1
pymongo==4.6.2
sounddevice==0.4.6
pyaudio==0.2.14
SpeechRecognition==3.10.0
tf-keras==2.18.0
plotly==5.24.1